refactor: Screen 오버로딩
