+++ app/src/main/java/nextstep/signup/ui/auth/screen/SignUpScreen.kt
@@ -5,18 +5,26 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Surface
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
+import kotlinx.coroutines.launch
+import nextstep.signup.R
 import nextstep.signup.ui.auth.component.SignUpConfirmButton
 import nextstep.signup.ui.auth.component.SignUpForm
 import nextstep.signup.ui.auth.component.SignUpTitle
@@ -26,75 +34,113 @@ import nextstep.signup.ui.interaction.clearFocusWith
 import nextstep.signup.ui.theme.SignupTheme
 
 @Composable
-fun SignUpScreen(
+fun SignUpScreen() {
+    var formState by rememberSaveable {
+        mutableStateOf(SignUpFormState("", "", "", ""))
+    }
+    val onSignUpFormStateChange: (SignUpFormState) -> Unit = { new: SignUpFormState ->
+        formState = new
+    }
+    val onEmailChange = { newEmail: String ->
+        onSignUpFormStateChange(formState.copy(email = newEmail))
+    }
+    val onUserNameChange = { newUserName: String ->
+        onSignUpFormStateChange(formState.copy(userName = newUserName))
+    }
+    val onPasswordChange = { newPassword: String ->
+        onSignUpFormStateChange(formState.copy(password = newPassword))
+    }
+    val onPasswordConfirmChange = { newPasswordConfirm: String ->
+        onSignUpFormStateChange(formState.copy(passwordConfirm = newPasswordConfirm))
+    }
+    val enableSignUp by remember {
+        derivedStateOf {
+            formState.enableSignUp
+        }
+    }
+    val snackbarHostState = SnackbarHostState()
+    val coroutineScope = rememberCoroutineScope()
+    Scaffold(
+        snackbarHost = { SnackbarHost(snackbarHostState) }
+    ) { innerPadding ->
+        val completeMessage = stringResource(id = R.string.sign_up_success)
+        val onDone: () -> Unit = remember {
+            {
+                coroutineScope.launch {
+                    snackbarHostState.showSnackbar(completeMessage)
+                }
+            }
+        }
+        SignUpScreen(
+            modifier = Modifier.padding(innerPadding),
+            signUpFormState = formState,
+            onUserNameChange = onUserNameChange,
+            onEmailChange = onEmailChange,
+            onPasswordChange = onPasswordChange,
+            onPasswordConfirmChange = onPasswordConfirmChange,
+            enableSignUp = enableSignUp,
+            onDoneSignUp = onDone
+        )
+    }
+}
+
+@Composable
+internal fun SignUpScreen(
     modifier: Modifier = Modifier,
     signUpFormState: SignUpFormState,
     onUserNameChange: (String) -> Unit,
-    onEmailChange: (String) ->  Unit,
+    onEmailChange: (String) -> Unit,
     onPasswordChange: (String) -> Unit,
     onPasswordConfirmChange: (String) -> Unit,
     enableSignUp: Boolean,
     onDoneSignUp: () -> Unit
 ) {
-    Surface {
-        Column(
-            modifier = modifier
-                .fillMaxSize()
-                .clearFocusWith(LocalFocusManager.current)
-                .padding(top = 30.dp)
-                .padding(horizontal = 36.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            SignUpTitle()
-            Spacer(modifier = Modifier.height(25.dp))
-            SignUpForm(
-                signUpFormState = signUpFormState,
-                onUserNameChange = onUserNameChange,
-                onEmailChange = onEmailChange,
-                onPasswordChange = onPasswordChange,
-                onPasswordConfirmChange = onPasswordConfirmChange,
-                onDoneSignUp = onDoneSignUp
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-            SignUpConfirmButton(onDoneSignUp, enableSignUp)
-        }
-    }
-}
 
-@Preview(showSystemUi = true)
-@Composable
-private fun Preview(
-    @PreviewParameter(SignUpPreviewParamsProvider::class) state: SignUpFormState
-) {
-    SignupTheme {
-        var signUpFormState by remember { mutableStateOf(state) }
-        val onSignUpFormStateChange: (SignUpFormState) -> Unit = { new ->
-            signUpFormState = new
-        }
-        val onUserNameChange: (String) -> Unit = { userName ->
-            onSignUpFormStateChange(signUpFormState.copy(userName = userName))
-        }
-
-        val onEmailChange: (String) -> Unit = { email ->
-            onSignUpFormStateChange(signUpFormState.copy(email = email))
-        }
-
-        val onPasswordChange: (String) -> Unit = { password ->
-            onSignUpFormStateChange(signUpFormState.copy(password = password))
-        }
-
-        val onPasswordConfirmChange: (String) -> Unit = { passwordConfirm ->
-            onSignUpFormStateChange(signUpFormState.copy(passwordConfirm = passwordConfirm))
-        }
-        val onDoneSignUp: () -> Unit = {}
-        SignUpScreen(
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .clearFocusWith(LocalFocusManager.current)
+            .padding(top = 30.dp)
+            .padding(horizontal = 36.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        SignUpTitle()
+        Spacer(modifier = Modifier.height(25.dp))
+        SignUpForm(
             signUpFormState = signUpFormState,
             onUserNameChange = onUserNameChange,
             onEmailChange = onEmailChange,
             onPasswordChange = onPasswordChange,
             onPasswordConfirmChange = onPasswordConfirmChange,
-            onDoneSignUp = onDoneSignUp,
-            enableSignUp = true
+            onDoneSignUp = onDoneSignUp
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+        SignUpConfirmButton(onDoneSignUp, enableSignUp)
+    }
+}
+
+@Preview(showSystemUi = true)
+@Composable
+private fun Preview1() {
+    SignupTheme {
+        SignUpScreen()
+    }
+}
+
+@Preview(showSystemUi = true)
+@Composable
+private fun Preview2(
+    @PreviewParameter(SignUpPreviewParamsProvider::class) state: SignUpFormState
+) {
+    SignupTheme {
+        SignUpScreen(
+            signUpFormState = state,
+            onUserNameChange = {},
+            onEmailChange = {},
+            onPasswordChange = {},
+            onPasswordConfirmChange = {},
+            onDoneSignUp = {},
+            enableSignUp = state.enableSignUp
         )
     }
 }
